LayoutTests/http/tests/workers/service/basic-register-expected.txt

CONSOLE MESSAGE: line 11: Registration failed with error: UnknownError: serviceWorker job scheduling is not yet implemented
